Fix parcel build
I had an issue with the API. I included the RemNoteAPI from multiple
places. This created multiple instances of the API, each with their own
listeners trying to answer the same message multiple times. I resolved
this by including it with a <script> and not import-ing it as an ES6
module.

Next I try to make the API a singleton.